requirements.txt
# finMapReduce - Comprehensive Requirements
# Python 3.10.18 compatible

# Core LLM Integration & API
langchain==0.3.26
langchain-openai==0.3.28
langchain-community==0.3.26
langchain-core==0.3.72
langchain-text-splitters==0.3.8
tiktoken==0.9.0
openai==1.88.0
anthropic==0.46.0

# FastAPI Web Application
fastapi==0.116.0
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-dotenv==1.1.1
starlette==0.46.2

# Document Processing & PDF Handling
pypdf==5.9.0
pdfminer.six==20231228
PyMuPDF==1.26.3
pymupdf4llm==0.0.25
unstructured==0.18.1
marker-pdf==1.8.2
markdownify==1.1.0
pdf2image==1.17.0
pdfplumber==0.11.7
camelot-py==1.0.0
tabula-py==2.10.0
python-docx==1.2.0

# Data Processing & Utilities
pyyaml==6.0.2
json5==0.12.0
tqdm==4.67.1
requests==2.32.4
numpy==1.26.4
pandas==2.3.0
scipy==1.15.3

# Machine Learning & Evaluation
torch==2.7.1
transformers==4.52.4
sentence-transformers==5.0.0
scikit-learn==1.7.0
bert-score==0.3.13
datasets==2.14.4
huggingface-hub==0.34.2

# Async Processing & Concurrency
aiofiles==24.1.0
aiohttp==3.12.13
aiosignal==1.3.2
anyio==4.9.0
httpx==0.28.1
httpcore==1.0.9

# Text Processing & NLP
nltk==3.9.1
textdistance==4.6.3
langdetect==1.0.9
Unidecode==1.4.0
python-Levenshtein==0.27.1

# Data Storage & Serialization
SQLAlchemy==2.0.41
pyarrow==20.0.0
orjson==3.10.18

# Rate Limiting & Client Management
pyrate-limiter==3.9.0
tenacity==8.5.0
backoff==2.2.1

# System Utilities
psutil==5.9.0
python-magic==0.4.27
filetype==1.2.0
rich==14.0.0
typer==0.16.0

# Testing & Development
pytest==8.4.1
pytest-asyncio
httpx>=0.25.0

# Optional: OCR and Advanced Document Processing
pytesseract==0.3.13
pillow==10.4.0
opencv-python==4.11.0.86

# Optional: Additional ML Models
timm==1.0.13
accelerate==1.8.1
safetensors==0.5.3

# Optional: Jupyter and Analysis
matplotlib==3.10.3
seaborn==0.13.2
jupyter-client
ipython